requirements.txt
django
djangorestframework
django_filter
#pymysql
mysqlclient
